refactor(mapview): extract types, presets and pure helpers to modules
Move the module-level prelude out of the 2630-line MapView component into
src/components/mapview/:

- types.ts    — shared domain types
- presets.ts  — TAIPEI_ROUTE_PRESETS route data
- helpers.ts  — pure geometry/GPX/sampling/fetch helpers

Behavior-preserving: the extracted code is byte-identical to the original
(verified by diff), only module-level declarations moved and imports were
rewired. The component body and all hooks are unchanged. tsc, eslint, the
vitest suite and next build all pass.
